+++ views/admin/motion-type/_policy.php
@@ -16,37 +16,6 @@ foreach (IPolicy::getPolicies() as $policy) {
     $policies[$policy::getPolicyID()] = $policy::getPolicyName();
 }
 
-$printUserGroupSelector = function (string $id, string $formName, ConsultationMotionType $motionType, IPolicy $currentPolicy) {
-    if (ConsultationUserGroup::consultationHasLoadableUserGroups($motionType->getConsultation())) {
-        $groupLoadUrl = \app\components\UrlHelper::createUrl('/admin/users/search-groups');
-    } else {
-        $groupLoadUrl = '';
-    }
-    if (is_a($currentPolicy, \app\models\policies\UserGroups::class)) {
-        $preselectedUserGroupsIds = array_map(function (ConsultationUserGroup $group): int { return $group->id; }, $currentPolicy->getAllowedUserGroups());
-    } else {
-        $preselectedUserGroupsIds = [];
-    }
-    ?>
-    <div class="userGroupSelect" data-load-url="<?= Html::encode($groupLoadUrl) ?>">
-        <select id="<?= $id ?>" name="type[<?= $formName ?>][groups][]" multiple
-                placeholder="<?= Yii::t('admin', 'motion_type_group_ph') ?>" title="<?= Yii::t('admin', 'motion_type_group_title') ?>">
-            <?php
-            foreach ($motionType->getConsultation()->getAllAvailableUserGroups($preselectedUserGroupsIds) as $group) {
-                echo '<option value="' . $group->id . '"';
-                if (is_a($currentPolicy, \app\models\policies\UserGroups::class) && $currentPolicy->allowsUserGroup($group)) {
-                    echo ' selected';
-                }
-                echo '>';
-                echo Html::encode($group->getNormalizedTitle());
-                echo '</option>';
-            }
-            ?>
-        </select>
-    </div>
-    <?php
-}
-
 ?>
 <h2 class="h3"><?= Yii::t('admin', 'motion_type_perm') ?></h2>
 
@@ -65,7 +34,7 @@ $printUserGroupSelector = function (string $id, string $formName, ConsultationMo
             $policies,
             ['id' => 'typePolicyMotions', 'class' => 'stdDropdown policySelect']
         );
-        $printUserGroupSelector('typePolicyMotionsGroups', 'policyMotions', $motionType, $currentPolicy);
+        echo $this->render('_usergroup_selector', ['id' => 'typePolicyMotionsGroups', 'formName' => 'policyMotions', 'motionType' => $motionType, 'currentPolicy' => $currentPolicy]);
         ?>
     </div>
 </div>
@@ -86,7 +55,7 @@ $printUserGroupSelector = function (string $id, string $formName, ConsultationMo
             $policies,
             ['id' => 'typePolicySupportMotions', 'class' => 'stdDropdown policySelect']
         );
-        $printUserGroupSelector('typePolicySupportMotionsGroups', 'policySupportMotions', $motionType, $currentPolicy);
+        echo $this->render('_usergroup_selector', ['id' => 'typePolicySupportMotionsGroups', 'formName' => 'policySupportMotions', 'motionType' => $motionType, 'currentPolicy' => $currentPolicy]);
         ?>
     </div>
 </div>
@@ -144,7 +113,7 @@ $printUserGroupSelector = function (string $id, string $formName, ConsultationMo
             $policies,
             ['id' => 'typePolicyAmendments', 'class' => 'stdDropdown policySelect']
         );
-        $printUserGroupSelector('typePolicyAmendmentsGroups', 'policyAmendments', $motionType, $currentPolicy);
+        echo $this->render('_usergroup_selector', ['id' => 'typePolicyAmendmentsGroups', 'formName' => 'policyAmendments', 'motionType' => $motionType, 'currentPolicy' => $currentPolicy]);
         ?>
     </div>
 </div>
@@ -199,7 +168,7 @@ $printUserGroupSelector = function (string $id, string $formName, ConsultationMo
             $policies,
             ['id' => 'typePolicySupportAmendments', 'class' => 'stdDropdown policySelect']
         );
-        $printUserGroupSelector('typePolicySupportAmendmentsGroups', 'policySupportAmendments', $motionType, $currentPolicy);
+        echo $this->render('_usergroup_selector', ['id' => 'typePolicySupportAmendmentsGroups', 'formName' => 'policySupportAmendments', 'motionType' => $motionType, 'currentPolicy' => $currentPolicy]);
         ?>
     </div>
 </div>
@@ -258,7 +227,7 @@ $printUserGroupSelector = function (string $id, string $formName, ConsultationMo
             $policies,
             ['id' => 'typePolicyComments', 'class' => 'stdDropdown policySelect']
         );
-        $printUserGroupSelector('typePolicyCommentsGroups', 'policyComments', $motionType, $currentPolicy);
+        echo $this->render('_usergroup_selector', ['id' => 'typePolicyCommentsGroups', 'formName' => 'policyComments', 'motionType' => $motionType, 'currentPolicy' => $currentPolicy]);
         ?>
     </div>
 </div>